+++ R/board_connect.R
@@ -30,16 +30,8 @@
 #' ))
 #' board %>% pin_read("numbers")
 #' ```
-#' 
-#' You can find the URL of a pin with [pin_browse()].
 #'
-#' # Local caching for Posit Connect
-#' 
-#' The pins package maintains local per-session caches for _users_ and _content_ 
-#' from your Connect server. If your cache gets into a bad state (for example, 
-#' user names have changed on the server or a pin was deleted on the server, but 
-#' your local machine doesn't know about the change yet), you can clear your 
-#' local cache by restarting your R session.
+#' You can find the URL of a pin with [pin_browse()].
 #'
 #' @inheritParams new_board
 #' @inheritParams board_url
@@ -133,14 +125,9 @@ board_rsconnect <- function(auth = c("auto", "manual", "envvar", "rsconnect"),
                             cache = NULL,
                             name = "posit-connect",
                             versioned = TRUE,
-                            use_cache_on_failure = is_interactive(),
-                            versions = deprecated()) {
+                            use_cache_on_failure = is_interactive()) {
 
-  lifecycle::deprecate_warn("1.1.0", "board_rsconnect()", "board_connect()")
-  if (lifecycle::is_present(versions)) {
-    lifecycle::deprecate_stop("1.0.0", "board_rsconnect(versions)", "board_rsconnect(versioned)")
-    versioned <- versions
-  }
+  lifecycle::deprecate_stop("1.1.0", "board_rsconnect()", "board_connect()")
 
   board_connect(
     auth = auth,
@@ -437,21 +424,9 @@ board_pin_find.pins_board_connect <- function(board,
 
 # Content -----------------------------------------------------------------
 
-the <- rlang::new_environment()
-the$connect_content_cache <- rlang::new_environment()
-the$connect_user_cache <- rlang::new_environment()
-
 rsc_content_find <- function(board, name, version = NULL, warn = TRUE) {
-  name <- rsc_parse_name(name)
-  content <- rlang::env_cache(
-    env = the$connect_content_cache,
-    nm = name$full %||% name$name,
-    default = rsc_content_find_live(board, name, version = NULL, warn = TRUE)
-  )
-  content
-}
 
-rsc_content_find_live <- function(board, name, version = NULL, warn = TRUE) {
+  name <- rsc_parse_name(name)
 
   # https://docs.rstudio.com/connect/api/#get-/v1/content
   json <- rsc_GET(board, "v1/content", list(name = name$name))
@@ -471,7 +446,7 @@ rsc_content_find_live <- function(board, name, version = NULL, warn = TRUE) {
     name$full <- paste0(owner, "/", name$name)
 
     if (warn) {
-      cli::cli_alert_warning("Please use full name when reading a pin: {.val {name$full}}, not {.val {name$name}}.")
+      cli::cli_alert_warning("Use a fully specified name including user name: {.val {name$full}}, not {.val {name$name}}.")
     }
     selected <- json[[1]]
   } else {
@@ -573,7 +548,6 @@ rsc_content_version_cached <- function(board, guid) {
 rsc_content_delete <- function(board, name) {
   content <- rsc_content_find(board, name)
   rsc_DELETE(board, rsc_v1("content", content$guid))
-  env_unbind(the$connect_content_cache, name)
   invisible(NULL)
 }
 
@@ -588,11 +562,7 @@ rsc_parse_name <- function(x) {
 }
 
 rsc_user_name <- function(board, guid) {
-  rlang::env_cache(
-    env = the$connect_user_cache,
-    nm = guid,
-    rsc_GET(board, rsc_v1("users", guid))$username
-  )
+  rsc_GET(board, rsc_v1("users", guid))$username
 }
 
 # helpers -----------------------------------------------------------------