+++ R/pin-read-write.R
@@ -64,6 +64,8 @@ pin_read <- function(board, name, version = NULL, hash = NULL, ...) {
 #'   use the default for `board`
 #' @param tags A character vector of tags for the pin; most important for
 #'   discoverability on shared boards.
+#' @param urls A character vector of URLs for more info on the pin, such as a
+#'   link to a wiki or other documentation.
 #' @param force_identical_write Store the pin even if the pin contents are
 #'   identical to the last version (compared using the hash). Only the pin
 #'   contents are compared, not the pin metadata. Defaults to `FALSE`.
@@ -71,15 +73,20 @@ pin_read <- function(board, name, version = NULL, hash = NULL, ...) {
 #' @export
 pin_write <- function(board, x,
                       name = NULL,
+                      ...,
                       type = NULL,
                       title = NULL,
                       description = NULL,
                       metadata = NULL,
                       versioned = NULL,
                       tags = NULL,
-                      ...,
+                      urls = NULL,
                       force_identical_write = FALSE) {
   check_board(board, "pin_write", "pin")
+  dots <- list2(...)
+  if (!missing(...) && (is.null(names(dots)) || names(dots)[[1]] == "")) {
+    cli::cli_abort('Arguments after the dots `...` must be named, like {.code type = "json"}.')
+  }
 
   if (is.null(name)) {
     name <- enexpr(x)
@@ -92,6 +99,7 @@ pin_write <- function(board, x,
   }
   check_metadata(metadata)
   check_character(tags, allow_null = TRUE)
+  check_character(urls, allow_null = TRUE)
   if (!is_string(name)) {
     abort("`name` must be a string")
   }
@@ -110,7 +118,8 @@ pin_write <- function(board, x,
     type = type,
     title = title %||% default_title(name, data = x),
     description = description,
-    tags = tags
+    tags = tags,
+    urls = urls
   )
   meta$user <- metadata